refactor(variant): move buildVariantSql + VariantScope to variant-scope.ts (BLD-1086 Phase 0a)
- Extract buildVariantSql, variantDrizzleConditions, VariantScope from exercise-history.ts
  to new lib/db/variant-scope.ts
- Extend VariantScope with gripType and stackUnitAtLog optional fields
  (existing call-sites pass undefined → no behavior change)
- Re-export from exercise-history.ts for backward compatibility
- All existing tests must remain unchanged
